Move simulator to utils
