+++ dogm/demo/utils/include/simulator.h
@@ -1,0 +1,67 @@
+/*
+MIT License
+
+Copyright (c) 2019 Michael Kösel
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+*/
+
+#ifndef SIMULATOR_H
+#define SIMULATOR_H
+
+#include <glm/glm.hpp>
+#include <vector>
+
+struct Vehicle
+{
+    Vehicle(const int width, const glm::vec2& pos, const glm::vec2& vel) : width(width), pos(pos), vel(vel) {}
+
+    void move(float dt) { pos += vel * dt; }
+
+    int width;
+    glm::vec2 pos;
+    glm::vec2 vel;
+};
+
+struct SimulationStep
+{
+    std::vector<Vehicle> vehicles;
+    std::vector<float> measurements;
+};
+
+using SimulationData = std::vector<SimulationStep>;
+
+struct Simulator
+{
+    Simulator(int _num_horizontal_scan_points, const float _field_of_view)
+        : num_horizontal_scan_points(_num_horizontal_scan_points), field_of_view(_field_of_view)
+    {
+    }
+
+    void addVehicle(const Vehicle& vehicle) { vehicles.push_back(vehicle); }
+
+    SimulationData update(int steps, float dt);
+    void addVehicleDetectionsToMeasurement(const Vehicle& vehicle, std::vector<float>& measurement) const;
+
+    int num_horizontal_scan_points;
+    float field_of_view;
+    std::vector<Vehicle> vehicles;
+};
+
+#endif  // SIMULATOR_H
